Revert "Fix Stepper performance bug (#97354)" (#97429)
This reverts commit 8c0021bd1986e49e9b9931fd23a50e3ad9cc0b1f.
